+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/SchoolReader.kt
@@ -1,5 +1,7 @@
 package com.sexysisters.tojserverv2.domain.school
 
+import com.sexysisters.tojserverv2.domain.school.domain.School
+
 interface SchoolReader {
     fun getSchool(code: String): School
 }
