+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/SchoolStore.kt
@@ -1,5 +1,7 @@
 package com.sexysisters.tojserverv2.domain.school
 
+import com.sexysisters.tojserverv2.domain.school.domain.School
+
 interface SchoolStore {
     fun store(school: School): School
 }
