+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/policy/InfoRangePolicy.kt
@@ -1,8 +1,8 @@
 package com.sexysisters.tojserverv2.domain.school.policy
 
-import com.sexysisters.tojserverv2.domain.school.School
+import com.sexysisters.tojserverv2.domain.school.domain.Division
+import com.sexysisters.tojserverv2.domain.school.domain.School
 import com.sexysisters.tojserverv2.domain.school.exception.SchoolException
-import com.sexysisters.tojserverv2.domain.school.type.Division
 import com.sexysisters.tojserverv2.domain.student.domain.Student
 import org.springframework.core.annotation.Order
 import org.springframework.stereotype.Component
