+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/policy/SchoolPolicy.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.domain.school.policy
 
-import com.sexysisters.tojserverv2.domain.school.School
+import com.sexysisters.tojserverv2.domain.school.domain.School
 import com.sexysisters.tojserverv2.domain.student.domain.Student
 
 interface SchoolPolicy {
