+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/school/service/SchoolServiceImpl.kt
@@ -6,6 +6,7 @@ import com.sexysisters.tojserverv2.domain.school.SchoolReader
 import com.sexysisters.tojserverv2.domain.school.SchoolStore
 import com.sexysisters.tojserverv2.domain.school.policy.SchoolPolicy
 import com.sexysisters.tojserverv2.domain.student.StudentReader
+import com.sexysisters.tojserverv2.domain.student.domain.makeRelation
 import com.sexysisters.tojserverv2.infrastructure.neis.NeisSchoolReader
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -29,7 +30,7 @@ class SchoolServiceImpl(
     @Transactional
     override fun createSchool(code: String): String {
         val initSchool = neisSchoolReader.getSchoolByCode(code)
-        return schoolStore.store(initSchool).code
+        return schoolStore.store(initSchool).code.value
     }
 
     @Transactional
@@ -37,7 +38,7 @@ class SchoolServiceImpl(
         val student = studentReader.getCurrentStudent()
         val school = schoolReader.getSchool(code)
         schoolPolicy.forEach { it.check(student, school) }
-        school.makeRelation(student)
+        student.makeRelation(school)
         return student.status.description
     }
 }
