+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/wiki/WikiReader.kt
@@ -1,5 +1,7 @@
 package com.sexysisters.tojserverv2.domain.wiki
 
+import com.sexysisters.tojserverv2.domain.wiki.domain.Wiki
+
 interface WikiReader {
     fun getWiki(id: Long): Wiki
 }
