+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/wiki/WikiStore.kt
@@ -1,5 +1,7 @@
 package com.sexysisters.tojserverv2.domain.wiki
 
+import com.sexysisters.tojserverv2.domain.wiki.domain.Wiki
+
 interface WikiStore {
     fun store(wiki: Wiki): Wiki
 }
