+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/wiki/policy/UpdateAuthority.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.domain.wiki.policy
 
-import com.sexysisters.tojserverv2.domain.school.School
+import com.sexysisters.tojserverv2.domain.school.domain.School
 import com.sexysisters.tojserverv2.domain.student.domain.Student
 import com.sexysisters.tojserverv2.domain.student.exception.StudentException
 import org.springframework.core.annotation.Order
