+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/school/SchoolRepository.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.infrastructure.school
 
-import com.sexysisters.tojserverv2.domain.school.School
+import com.sexysisters.tojserverv2.domain.school.domain.School
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface SchoolRepository : JpaRepository<School, Long> {
