+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/student/StudentCustomRepository.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.infrastructure.student
 
-import com.sexysisters.tojserverv2.domain.school.School
+import com.sexysisters.tojserverv2.domain.school.domain.School
 
 interface StudentCustomRepository {
     fun checkAlreadyExists(
