+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/student/StudentCustomRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.sexysisters.tojserverv2.infrastructure.student
 
 import com.querydsl.jpa.impl.JPAQueryFactory
-import com.sexysisters.tojserverv2.domain.school.School
+import com.sexysisters.tojserverv2.domain.school.domain.School
 import com.sexysisters.tojserverv2.domain.student.domain.QStudent.student
 import org.springframework.stereotype.Component
