+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/student/StudentRepository.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.infrastructure.student
 
-import com.sexysisters.tojserverv2.domain.school.School
+import com.sexysisters.tojserverv2.domain.school.domain.School
 import com.sexysisters.tojserverv2.domain.student.domain.Status
 import com.sexysisters.tojserverv2.domain.student.domain.Student
 import org.springframework.data.jpa.repository.JpaRepository
