+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/wiki/WikiReaderImpl.kt
@@ -1,7 +1,7 @@
 package com.sexysisters.tojserverv2.infrastructure.wiki
 
-import com.sexysisters.tojserverv2.domain.wiki.Wiki
 import com.sexysisters.tojserverv2.domain.wiki.WikiReader
+import com.sexysisters.tojserverv2.domain.wiki.domain.Wiki
 import com.sexysisters.tojserverv2.domain.wiki.exception.WikiException
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
