+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/wiki/WikiRepository.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.infrastructure.wiki
 
-import com.sexysisters.tojserverv2.domain.wiki.Wiki
+import com.sexysisters.tojserverv2.domain.wiki.domain.Wiki
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface WikiRepository : JpaRepository<Wiki, Long>
