+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/wiki/WikiStoreImpl.kt
@@ -1,6 +1,6 @@
 package com.sexysisters.tojserverv2.infrastructure.wiki
 
-import com.sexysisters.tojserverv2.domain.wiki.Wiki
+import com.sexysisters.tojserverv2.domain.wiki.domain.Wiki
 import com.sexysisters.tojserverv2.domain.wiki.WikiStore
 import org.springframework.stereotype.Component
